requirements.txt
streamlit
pandas
python-docx
PyPDF2
textblob
google-generativeai
json-repair